Create CurrencyCreateInput and UpdateInput model
Add two new models to allow for creation and updating of currency data
in postgres table.
