--- src/main/java/com/mikejacks/international_currency_converter/localization/model/Currency.java
@@ -1,39 +1,0 @@
-package com.mikejacks.international_currency_converter.localization.model;
-
-public class Currency {
-    private String baseCode;
-    private String targetCode;
-    private Double conversionRate;
-
-    public Currency() {}
-
-    public Currency(String baseCode, String targetCode, Double conversionRate) {
-        this.setBaseCode(baseCode);
-        this.setTargetCode(targetCode);
-        this.setConversionRate(conversionRate);
-    }
-
-    public String getBaseCode() {
-        return baseCode;
-    }
-
-    public void setBaseCode(String baseCode) {
-        this.baseCode = baseCode;
-    }
-
-    public String getTargetCode() {
-        return targetCode;
-    }
-
-    public void setTargetCode(String targetCode) {
-        this.targetCode = targetCode;
-    }
-
-    public Double getConversionRate() {
-        return conversionRate;
-    }
-
-    public void setConversionRate(Double conversionRate) {
-        this.conversionRate = conversionRate;
-    }
-}
